Got it finished, tested it works fine for all scenarios for me.
it's still in the older version of the project and I won't have time to merge it for tomorrow before I meet you in the afternoon.

I left some changes for reading the combobox of tickets instead of strings that are commented out in tokenmachine gui in the createTicket and createTimedTicket functions and the class instantiation.

There are also some lines you should comment out in the SetupFile function after running it once or tickets will be overwritten each time you use the token machine gui and every stations ticketlist will be written to file.

going to bed now.

Gareth